Attempt to breakout Protocol implementations into their own module
